+++ Tests/KeyboardKitTests/Extensions/UITextDocumentProxy+CurrentWordTests.swift
@@ -114,6 +114,27 @@ class UITextDocumentProxy_CurrentWordTests: QuickSpec {
             }
         }
         
+        describe("is cursor at the beginning of a new sentence") {
+            
+            func result(for preCursorPart: String) -> Bool {
+                proxy.documentContextBeforeInput = preCursorPart
+                return proxy.isCursorAtTheBeginningOfASentence
+            }
+            
+            it("returns true if pre cursor part is missing") {
+                expect(proxy.isCursorAtTheBeginningOfASentence).to(beTrue())
+            }
+            
+            it("returns false if pre cursor part ends with a non-word delimiter") {
+                expect(result(for: "foo")).to(beFalse())
+            }
+            
+            it("returns false if pre cursor part ends with a word delimiter") {
+                expect(result(for: "foo.")).to(beTrue())
+                expect(result(for: "foo! ")).to(beTrue())
+            }
+        }
+        
         describe("is cursor at the end of the current word") {
             
             it("returns false if current word is missing") {